+++ CHANGELOG.md
@@ -27,3 +27,4 @@
     * Object.spawn
     * Object.add_callback
 * Item.find, Item.find_all : Now work with custom RMT Objects
+* Added Instance.number